added new method too check if employee exist

+++ src/main/java/com/rest/micronaut/controller/EmployeeController.java
@@ -43,7 +43,10 @@ public class EmployeeController {
 
 	@Delete("/{id}")
 	public void deleteEmployeeDetails(Long id) {
-		employeeService.deleteEmployeeDetails(id);
+		if(employeeService.isExist(id)) {
+			employeeService.deleteEmployeeDetails(id);
+		}
+		return null;
 	}
 
 }
